Add portainer as service
Portainer spins up a webserver where we can manage containers running on the system. Useful if we for example want to restart a container from a location where ssh's wouldnt be optimal. E.g outside of LAN.

Since portainer dos'nt store and critical data in its volume, we can store it in a named volume. There is no need to binding it to a location on the host.

https://www.portainer.io/