Proxy part 1, network
we'll call the network proxy since it will include all services that should be exposed on the proxy network

Bridge will make it so that containers within the proxy network wont be able to communicate with containers that are not in the proxy network.

external false, we expect docker to create this network for us when starting the containers for the first time.

subnet. We want to assign the proxy network to a specific ip range. Some services (homeassistant) want to know the ip of the reverse proxy. Otherwise, it will block all requests. I presume this is to mitigate man in the middle attacks. More on the ha config later. 10.1.0.0/16 means that containers within this network can have an ip of 10.1.0.0 to 10.1.255.255 (65k containers). where 10.1 defines the network itself.