Proxy part 2, traefik service
We'll use traefik as our reverse proxy. Mainly since it goes very well hand in hand with docker-compose. It also puts as much of our config in the docker-compose file. Which i personally like, since i'll have fewer places to look when changing the infra of the server.

We'll bind port 80 and 443 to the host to listen for http and https requests to the server.

We also make sure to store the ssl certificates on the host so that we can reuse them if the container would be removed. Since letsencrypt enforce rate-limit.

Commands
I added some comments above the commands to describe what they do. Traefik has good documentation about this.