Proxy part 5, portainer behind proxy
Adding new services behind the reverse proxy with new sub-domains is now very straight forward. we'll start with portainer.

we change the "ports" to "expose" since we dont want/need the container to post its port on the host network, since it will be proxied by traefik.

we'll then add the container to the `proxy` network.

and finally add the labels so that traefik will pick them up and start proxying the traffic.